Add files via upload
This is my ongoing research paper based on security breaches of remote working. It tells the global human impact on human kind for working remotely in both positive and negative aspects. Moreover, I have added the analysis of data and research which has already done for this and with the help of that I made effort to research more about the topic and what are the new things which can be added and analysed with the help of the earlier data.